Backend/py_backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
sqlmodel>=0.0.8
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
httpx>=0.23.0

